refactor: extract destination-NAT compile into its own file (#4846)
Move compileNATDestination and its DNAT pool/port helpers
(parseDNATPoolAddress, appendDNATPortRange, parseDNATPortList) out of
compiler_nat.go into a new compiler_nat_destination.go in the same
package.

Pure code motion: bodies moved byte-for-byte, no signature or logic
change. The new file imports only fmt and strconv (the subset the moved
text uses); compiler_nat.go keeps the imports its remaining source and
static code needs.
